atm-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose==3.3.0
pydantic==2.11.3
asyncpg
dotenv 
